Added ResponseEntity on REST API, create errors information, add integration with Swagger

+++ src/main/java/com/example/demo/user/api/UserApiController.java
@@ -1,17 +1,20 @@
 package com.example.demo.user.api;
 
+import com.example.demo.error.EntityNotFoundException;
 import com.example.demo.user.User;
 import com.example.demo.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping
+@RequestMapping("/api")
 public class UserApiController {
     private UserRepository userRepository;
-
+    HttpStatus status;
     @Autowired
     public UserApiController(UserRepository userRepository){
         this.userRepository = userRepository;
@@ -22,18 +25,32 @@ public class UserApiController {
         return userRepository.findAll();
     }
 
-    @RequestMapping(value = "/users", method = RequestMethod.POST)
-    public User createUser(@RequestBody User user){
-        return userRepository.save(user);
-    }
+   @RequestMapping(value = "/users", method = RequestMethod.POST)
+    public ResponseEntity<User> createUser(@RequestBody User user){
 
-    @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
-    public User updateUser(@PathVariable String email, @RequestBody User user){
-        return userRepository.save(email, user);
-    }
+       if(!userRepository.exists(user.getEmail()))
+        {
+            status = HttpStatus.CREATED;
+        }
+        User saved = userRepository.save(user);
+        return new ResponseEntity<>(saved, status);
+   }
 
-    @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String email){
-        userRepository.delete(email);
-    }
+   @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
+        public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user) throws EntityNotFoundException {
+            if(!userRepository.exists(user.getEmail())){
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+       }
+       User saved = userRepository.update(email, user);
+       return new ResponseEntity<>(saved, HttpStatus.CREATED);
+   }
+
+   @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
+    public ResponseEntity<User> deleteUser(@PathVariable String email) throws EntityNotFoundException {
+            if(!userRepository.exists(email)){
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            userRepository.delete(email);
+            return new ResponseEntity<>(HttpStatus.OK);
+   }
 }
